Change grid charging to select entity and fixes
